requirements.txt
fastapi==0.111.1
fastapi-cli==0.0.4
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
uvicorn==0.30.1
wn==0.9.5